Lab 2: Fix bootloader running directly from SD card
Before this, my bootloader wouldn't work when running directly from the
SD card, but was fine when sent to a different bootloader (dwelch67's
bootloader05). After a ton of debugging, it turns out there's a single
byte already in the RX queue when booting. I have no idea what causes
that, but ignore the extra byte for now.